refactor: deleted all the tests
