--- tests/database/docker-compose.yml
@@ -1,9 +1,0 @@
-version: '3'
-
-services:
-  db:
-    image: postgres
-    env_file:
-      - ./.env
-    ports:
-      - 5432:5432
