--- tests/unit/conftest.py
@@ -1,33 +1,0 @@
-# -*- coding: utf-8 -*-
-import pytest
-
-from api import create_app, db
-
-
-@pytest.fixture()
-def app():
-    app = create_app("testing")
-    with app.app_context():
-        db.drop_all()
-        db.create_all()
-        yield app
-
-
-@pytest.fixture()
-def dev_app():
-    app = create_app("development")
-    yield app
-
-
-@pytest.fixture()
-def client(app):
-    app = create_app("testing")
-    with app.app_context():
-        db.drop_all()
-        db.create_all()
-        yield app.test_client()
-
-
-@pytest.fixture()
-def runner(app):
-    return app.test_cli_runner()
